debug atlas

+++ alltrain/atlasUtils.py
@@ -81,8 +81,8 @@ def getHd95(pred, target):
         return -1
 
 def getMask(labels, i):
-    return (labels == i).float()
-    # return (labels == i).half()
+    # return (labels == i).float()
+    return (labels == i).half()
 
 # def getTCMask(labels):
 #     return ((labels != 0) * (labels != 2)).float() #We use multiplication as AND
